notion_sync/requirements.txt
notion-client>=2.0.0
markdown-it-py>=3.0.0
python-frontmatter>=1.0.0
linkify-it-py>=2.0.0
python-dotenv
